+++ services/assets/functions/s3/upload-media.ts
@@ -6,38 +6,50 @@ import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
 import httpHeaderNormalizer from "@middy/http-header-normalizer";
 import httpMultipartBodyParser from "@middy/http-multipart-body-parser";
 import type { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
-import { roleChaining, multipleFileUpload, normalizeFilePath, getMediaBucket } from "../../lib";
+import { multipleFileUpload, normalizeFilePath, getMediaBucket } from "../../lib";
+import { UseAssumeRole, readConfigFile } from '@vastly/utils';
 
 const { AWS_REGION } = process.env;
 
 const s3uploadMedia = async (event: APIGatewayProxyEvent): Promise<APIGatewayProxyResult> => {
   try {
     const file = event.body?.["file"];
-    const filePath = normalizeFilePath(event.body?.["filePath"]);
     const clientName = event.pathParameters?.clientName;
+    const filePath = normalizeFilePath(event.body?.["filePath"]);
+    const idToken = event.headers?.["authorization"] || readConfigFile().token?.idToken;
+
 
     if (!file || !clientName || !filePath) {
       return {
         statusCode: 400,
         body: JSON.stringify({
-          message: "Incorrect body on request",
+          location: "Incorrect body on request",
         }),
       };
     }
 
-    const { credentials } = await roleChaining();
+    const assumeRoleResult = await UseAssumeRole(idToken, clientName, "CrossAccountAssetServiceRole");
+    if ("error" in assumeRoleResult) {
+      return {
+        statusCode: 404,
+        body: JSON.stringify({
+          location: `${assumeRoleResult.error}`,
+        }),
+      };
+    }
+
     const s3Client = new S3Client([
       {
         region: AWS_REGION,
         credentials: {
-          accessKeyId: credentials?.AccessKeyId,
-          secretAccessKey: credentials?.SecretAccessKey,
-          sessionToken: credentials?.SessionToken,
-          expiration: credentials?.Expiration,
+          accessKeyId: assumeRoleResult.Credentials?.AccessKeyId,
+          secretAccessKey: assumeRoleResult.Credentials?.SecretAccessKey,
+          sessionToken: assumeRoleResult.Credentials?.SessionToken,
+          expiration: assumeRoleResult.Credentials?.Expiration,
         },
       },
     ]);
-    const { bucketName } = await getMediaBucket(s3Client, clientName, credentials);
+    const { bucketName, client } = await getMediaBucket(s3Client, clientName, assumeRoleResult.Credentials);
 
     if (file.length > 1) {
       return await multipleFileUpload(file, filePath, bucketName ?? "");
@@ -53,32 +65,19 @@ const s3uploadMedia = async (event: APIGatewayProxyEvent): Promise<APIGatewayPro
       Bucket: bucketName,
       RequestPayer: "requester",
       Tagging: `type=${detectedMime?.split("/")[0]}&format=${detectedMime?.split("/")[1]}`,
-      ACL: "public-read",
     };
 
     console.log(`Writing image to bucket: ${Key}`);
     console.log("------- File Contents --------\n", input);
-
-    const newS3Client = new S3Client([
-      {
-        region: AWS_REGION,
-        credentials: {
-          accessKeyId: credentials?.AccessKeyId,
-          secretAccessKey: credentials?.SecretAccessKey,
-          sessionToken: credentials?.SessionToken,
-          expiration: credentials?.Expiration,
-        },
-      },
-    ]);
     const command = new PutObjectCommand(input);
-    const response = await newS3Client.send(command);
+    const response = await client?.send(command);
     const url = `https://${bucketName}.s3.amazonaws.com/${Key}`;
 
     return {
       isBase64Encoded: false,
       statusCode: 200,
       body: JSON.stringify({
-        fileUrl: url,
+        location: url,
         ...response,
       }),
     };
@@ -87,7 +86,7 @@ const s3uploadMedia = async (event: APIGatewayProxyEvent): Promise<APIGatewayPro
     return {
       statusCode: 500,
       body: JSON.stringify({
-        message: err.message || "Failed to upload image",
+        location: err.message || "Failed to upload image",
       }),
     };
   }