+++ apps/website/CHANGELOG.md
@@ -1,5 +1,11 @@
 # website
 
+## 0.1.3
+
+### Patch Changes
+
+- @vastly/ui@0.4.4
+
 ## 0.1.2
 
 ### Patch Changes