+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vastly/test-utils
 
+## 0.1.5
+
+### Patch Changes
+
+- @vastly/ui@0.4.4
+
 ## 0.1.4
 
 ### Patch Changes