+++ services/deploy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # deploy-service
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @vastly/utils@0.4.2
+
 ## 0.2.2
 
 ### Patch Changes